Krylov solver clean-up, HYPRE husk.
* Cracked PETSc/HYPRE logic out of krylov_solver.c
* Started the HYPRE implementation.
* The PETSc solver now uses SHARED_LIBRARY_SUFFIX.
