+++ core/CMakeLists.txt
@@ -13,7 +13,8 @@ add_library(polymec_core polymec.c arch.c allocators.c logging.c timer.c memory_
             text_buffer.c serializer.c interpreter_register_default_functions.c 
             thread_pool.c string_utils.c file_utils.c array_utils.c 
             linear_algebra.c local_matrix.c block_diagonal_matrix.c
-            sparse_local_matrix.c dense_local_matrix krylov_solver.c 
+            sparse_local_matrix.c dense_local_matrix krylov_solver.c
+            petsc_krylov_solver.c hypre_krylov_solver.c
             silo_file.c exchanger.c)
 
 # polymec.c depends on the creation of polymec_version.h.
